requirements.txt
# Core
torch
torchvision
Pillow
numpy
opencv-python

# PDF processing
pypdfium2

# OCR Models (transformers from source for LightOnOCR-2 support)
# pip install git+https://github.com/huggingface/transformers
transformers
accelerate

# Handwriting Detection
ultralytics
huggingface_hub

# Colab extras (install separately on Colab)
# qwen_vl_utils
# bitsandbytes
# peft